Add reboot menu
This menu is now in libcot to further separate it from the
main recovery code. Eventually it will be generated based
on the device in question but right now I'm too fucking
lazy.

I've also added a commented explanation for cot/external.h
(which was necessary to add due to requiring a core
recovery function)

Change-Id: I6d60373a40f03c5e4fbde5fbdf8884fcd9ad7d98